backend/requirements.txt
# Django and REST Framework
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache and Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.8.1
django-celery-results==2.5.1

# Authentication
django-allauth==0.63.3
# Additional dependencies for social account providers
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests-oauthlib==2.0.0
cffi==1.17.1

# AI and Machine Learning
openai==1.55.3

# PDF Generation
reportlab==4.2.0

# Push Notifications
firebase-admin==6.5.0

# API Documentation
drf-spectacular==0.27.2

# Environment Variables
python-decouple==3.8

# WSGI Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Security and Cryptography
cryptography==43.0.1
PyJWT==2.9.0
pycparser==2.22

# Build dependencies for cryptography
setuptools==75.1.0
wheel==0.44.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0

# Payments
stripe==10.1.0

# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0

# Error Tracking & Monitoring
sentry-sdk==1.40.0

# GraphQL
graphene-django==3.1.5
django-graphql-jwt==0.4.0

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0